Moved ImageLoaderTask and changed FsViewerAdapter
