+++ app/src/main/java/net/gsantner/opoc/util/ImageLoaderTask.java
@@ -1,30 +1,19 @@
 /*#######################################################
  *
- *   Maintained by Gregor Santner, 2016-
+ *   Maintained by Gregor Santner, 2017-
  *   https://gsantner.net/
  *
- *   License of this file: GNU GPLv3 (Commercial upon request)
+ *   License of this file: Apache 2.0 (Commercial upon request)
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *     https://github.com/gsantner/opoc/#licensing
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
 #########################################################*/
-package net.gsantner.opoc.service;
+package net.gsantner.opoc.util;
 
 import android.content.Context;
 import android.graphics.Bitmap;
 import android.os.AsyncTask;
 import android.support.v4.content.ContextCompat;
-import android.util.Log;
 
 import net.gsantner.markor.R;
 import net.gsantner.markor.util.AppSettings;
